fix: 统一maxScore数据类型。在没有评分时也可以获取maxScore。

+++ src/main/java/org/sustech/orion/controller/students/SubmissionController.java
@@ -224,7 +224,7 @@ public class SubmissionController {
 
         GradeResponseDTO gradeResponseDTO = ConvertDTO.toGradeResponseDTO(submission.getGrade());
         // 添加作业的最高分信息，将Integer转换为Double
-        gradeResponseDTO.setMaxScore(submission.getAssignment().getMaxScore().doubleValue());
+        gradeResponseDTO.setMaxScore(submission.getAssignment().getMaxScore());
 
         return ResponseEntity.ok(gradeResponseDTO);
     }
